Allow system pre-prompt context in non-chat mode, hide context in chat mode since used for history
e.g. could add to chat=False case a context of:

<bot>: I am a helpful assistant that can generate detailed elaborate question from topics such as biology, physics, mathematics, psychology, or computer science.

And then one gives the instruction:

I am a person seeking assistance from a bot.  Please generate an elaborate question.

And that will force the bot to behave a certain way because it has already effectively responded that way.