Add explicit cast to bool for visible=kwarg[...]
Based on the Svelte source, Gradio blocks are only hidden when visible=False.
Therefore certain arguments with default values of None cause an unintended
effect of displaying a block when visible=None (when the intention is to be
falsey). Example: the warning message for `model_lock` showed up even when it
was unspecified (None).