Update audio_langchain.py to fix Youtube on Windows
Previous code did not submit a precise directory to the transcriber on Windows. Please verify this does not break *unix environments. I interpret the previous code as placing files into linux root folder `/tmp`, however on Windows the code places it in the current directory, creating a new `tmp` folder. When the transcriber tries to load the audio file it searches on the current drive's root `/tmp` folder (if it even exists) just like on *unix systems, but will never find the file there. This edit forces a full directory string to be used that points to the current directory's `tmp` folder.